le epic

+++ NewHorizons/Handlers/PlanetCreationHandler.cs
@@ -466,7 +466,9 @@ namespace NewHorizons.Handlers
 
             if (body.Config.HeightMap != null)
             {
-                var res = (int)(body.Config.HeightMap.resolution / 4);
+                // resolution = tris per face
+                // divide by 4 to account for all the way around the equator
+                var res = body.Config.HeightMap.resolution / 4;
                 HeightMapBuilder.Make(go, sector, body.Config.HeightMap, body.Mod, res);
             }
